refactor(core): Update GitDiffResult imports and restructure git handling modules
