+++ src/cli/actions/remoteAction.ts
@@ -3,8 +3,9 @@ import os from 'node:os';
 import path from 'node:path';
 import pc from 'picocolors';
 import { execGitShallowClone } from '../../core/git/gitCommand.js';
-import { getRemoteRefs, isGitInstalled } from '../../core/git/gitHandle.js';
+import { getRemoteRefs } from '../../core/git/gitRemoteHandle.js';
 import { parseRemoteValue } from '../../core/git/gitRemoteParse.js';
+import { isGitInstalled } from '../../core/git/gitRepositoryHandle.js';
 import { RepomixError } from '../../shared/errorHandle.js';
 import { logger } from '../../shared/logger.js';
 import { Spinner } from '../cliSpinner.js';
