+++ src/core/output/outputGeneratorTypes.ts
@@ -1,6 +1,6 @@
 import type { RepomixConfigMerged } from '../../config/configSchema.js';
 import type { ProcessedFile } from '../file/fileTypes.js';
-import type { GitDiffResult } from '../git/gitHandle.js';
+import type { GitDiffResult } from '../git/gitDiffHandle.js';
 
 export interface OutputGeneratorContext {
   generationDate: string;
