+++ src/core/security/validateFileSafety.ts
@@ -2,7 +2,7 @@ import type { RepomixConfigMerged } from '../../config/configSchema.js';
 import { logger } from '../../shared/logger.js';
 import type { RepomixProgressCallback } from '../../shared/types.js';
 import type { ProcessedFile, RawFile } from '../file/fileTypes.js';
-import type { GitDiffResult } from '../git/gitHandle.js';
+import type { GitDiffResult } from '../git/gitDiffHandle.js';
 import { filterOutUntrustedFiles } from './filterOutUntrustedFiles.js';
 import { type SuspiciousFileResult, runSecurityCheck } from './securityCheck.js';
