+++ tests/core/git/gitDiffHandle.test.ts
@@ -1,62 +1,13 @@
 import { beforeEach, describe, expect, test, vi } from 'vitest';
-import {
-  getFileChangeCount,
-  getGitDiffs,
-  getRemoteRefs,
-  getStagedDiff,
-  getWorkTreeDiff,
-  isGitInstalled,
-  isGitRepository,
-} from '../../../src/core/git/gitHandle.js';
+import { getGitDiffs, getStagedDiff, getWorkTreeDiff } from '../../../src/core/git/gitDiffHandle.js';
 import { logger } from '../../../src/shared/logger.js';
 import { createMockConfig } from '../../testing/testUtils.js';
 
 vi.mock('../../../src/shared/logger');
 
-describe('gitHandle', () => {
+describe('gitDiffHandle', () => {
   beforeEach(() => {
     vi.resetAllMocks();
-  });
-
-  describe('getFileChangeCount', () => {
-    test('should count file changes correctly', async () => {
-      const mockFilenames = ['file1.ts', 'file2.ts', 'file1.ts', 'file3.ts', 'file2.ts'];
-
-      const mockExecGitLogFilenames = vi.fn().mockResolvedValue(mockFilenames);
-
-      const result = await getFileChangeCount('/test/dir', 5, {
-        execGitLogFilenames: mockExecGitLogFilenames,
-      });
-
-      expect(result).toEqual({
-        'file1.ts': 2,
-        'file2.ts': 2,
-        'file3.ts': 1,
-      });
-      expect(mockExecGitLogFilenames).toHaveBeenCalledWith('/test/dir', 5);
-    });
-
-    test('should return empty object when git command fails', async () => {
-      const mockExecGitLogFilenames = vi.fn().mockRejectedValue(new Error('git command failed'));
-
-      const result = await getFileChangeCount('/test/dir', 5, {
-        execGitLogFilenames: mockExecGitLogFilenames,
-      });
-
-      expect(result).toEqual({});
-      expect(logger.trace).toHaveBeenCalledWith('Failed to get file change counts:', 'git command failed');
-    });
-
-    test('should handle empty git log output', async () => {
-      const mockExecGitLogFilenames = vi.fn().mockResolvedValue([]);
-
-      const result = await getFileChangeCount('/test/dir', 5, {
-        execGitLogFilenames: mockExecGitLogFilenames,
-      });
-
-      expect(result).toEqual({});
-      expect(mockExecGitLogFilenames).toHaveBeenCalledWith('/test/dir', 5);
-    });
   });
 
   describe('getWorkTreeDiff', () => {
@@ -119,116 +70,6 @@ describe('gitHandle', () => {
       expect(result).toBe(mockDiff);
       expect(mockIsGitRepository).toHaveBeenCalledWith('/test/dir');
       expect(mockExecGitDiff).toHaveBeenCalledWith('/test/dir', ['--cached']);
-    });
-  });
-
-  describe('isGitRepository', () => {
-    test('should return true when directory is a git repository', async () => {
-      const mockExecGitRevParse = vi.fn().mockResolvedValue('true');
-
-      const result = await isGitRepository('/test/dir', {
-        execGitRevParse: mockExecGitRevParse,
-      });
-
-      expect(result).toBe(true);
-      expect(mockExecGitRevParse).toHaveBeenCalledWith('/test/dir');
-    });
-
-    test('should return false when directory is not a git repository', async () => {
-      const mockExecGitRevParse = vi.fn().mockRejectedValue(new Error('Not a git repository'));
-
-      const result = await isGitRepository('/test/dir', {
-        execGitRevParse: mockExecGitRevParse,
-      });
-
-      expect(result).toBe(false);
-      expect(mockExecGitRevParse).toHaveBeenCalledWith('/test/dir');
-    });
-  });
-
-  describe('isGitInstalled', () => {
-    test('should return true when git is installed', async () => {
-      const mockExecGitVersion = vi.fn().mockResolvedValue('git version 2.34.1');
-
-      const result = await isGitInstalled({
-        execGitVersion: mockExecGitVersion,
-      });
-
-      expect(result).toBe(true);
-      expect(mockExecGitVersion).toHaveBeenCalled();
-    });
-
-    test('should return false when git command fails', async () => {
-      const mockExecGitVersion = vi.fn().mockRejectedValue(new Error('Command not found: git'));
-
-      const result = await isGitInstalled({
-        execGitVersion: mockExecGitVersion,
-      });
-
-      expect(result).toBe(false);
-      expect(mockExecGitVersion).toHaveBeenCalled();
-      expect(logger.trace).toHaveBeenCalledWith('Git is not installed:', 'Command not found: git');
-    });
-
-    test('should return false when git version output contains error', async () => {
-      const mockExecGitVersion = vi.fn().mockResolvedValue('error: git not found');
-
-      const result = await isGitInstalled({
-        execGitVersion: mockExecGitVersion,
-      });
-
-      expect(result).toBe(false);
-      expect(mockExecGitVersion).toHaveBeenCalled();
-    });
-  });
-
-  describe('getRemoteRefs', () => {
-    test('should return refs when URL is valid', async () => {
-      const mockOutput = `
-a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6\trefs/heads/main
-b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7\trefs/heads/develop
-c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7r8\trefs/tags/v1.0.0
-`.trim();
-      const mockExecLsRemote = vi.fn().mockResolvedValue(mockOutput);
-
-      const result = await getRemoteRefs('https://github.com/user/repo.git', {
-        execLsRemote: mockExecLsRemote,
-      });
-
-      expect(result).toEqual(['main', 'develop', 'v1.0.0']);
-      expect(mockExecLsRemote).toHaveBeenCalledWith('https://github.com/user/repo.git');
-    });
-
-    test('should throw error when URL does not start with git@ or https://', async () => {
-      const mockExecLsRemote = vi.fn();
-
-      await expect(getRemoteRefs('invalid-url', { execLsRemote: mockExecLsRemote })).rejects.toThrow(
-        "Invalid URL protocol for 'invalid-url'. URL must start with 'git@' or 'https://'",
-      );
-
-      expect(mockExecLsRemote).not.toHaveBeenCalled();
-    });
-
-    test('should throw error when URL contains dangerous parameters', async () => {
-      const mockExecLsRemote = vi.fn();
-
-      await expect(
-        getRemoteRefs('https://github.com/user/repo.git --upload-pack=evil-command', {
-          execLsRemote: mockExecLsRemote,
-        }),
-      ).rejects.toThrow('Invalid repository URL. URL contains potentially dangerous parameters');
-
-      expect(mockExecLsRemote).not.toHaveBeenCalled();
-    });
-
-    test('should throw error when git command fails', async () => {
-      const mockExecLsRemote = vi.fn().mockRejectedValue(new Error('git command failed'));
-
-      await expect(
-        getRemoteRefs('https://github.com/user/repo.git', { execLsRemote: mockExecLsRemote }),
-      ).rejects.toThrow('Failed to get remote refs: git command failed');
-
-      expect(mockExecLsRemote).toHaveBeenCalledWith('https://github.com/user/repo.git');
     });
   });
 
@@ -308,6 +149,7 @@ c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7r8\trefs/tags/v1.0.0
         workTreeDiffContent: mockWorkTreeDiff,
         stagedDiffContent: mockStagedDiff,
       });
+      // createMockConfig sets cwd to the actual working directory, so we check the actual config value
       expect(mockGetWorkTreeDiff).toHaveBeenCalledWith(mockConfig.cwd);
       expect(mockGetStagedDiff).toHaveBeenCalledWith(mockConfig.cwd);
     });
