+++ tests/integration-tests/packager.test.ts
@@ -11,7 +11,7 @@ import type { ProcessedFile } from '../../src/core/file/fileTypes.js';
 import type { FileCollectTask } from '../../src/core/file/workers/fileCollectWorker.js';
 import fileCollectWorker from '../../src/core/file/workers/fileCollectWorker.js';
 import fileProcessWorker from '../../src/core/file/workers/fileProcessWorker.js';
-import type { GitDiffResult } from '../../src/core/git/gitHandle.js';
+import type { GitDiffResult } from '../../src/core/git/gitDiffHandle.js';
 import { generateOutput } from '../../src/core/output/outputGenerate.js';
 import { pack } from '../../src/core/packager.js';
 import { copyToClipboardIfEnabled } from '../../src/core/packager/copyToClipboardIfEnabled.js';
